added default parser settings

+++ api/classes/models.py
@@ -16,7 +16,7 @@ class Assignment(models.Model):
     description = models.TextField(blank=True)
     time_created = models.DateTimeField(auto_now=True)
     submit_by = models.DateTimeField(blank=True,null=True)
-    max_marks = models.DecimalField(max_digits=5,decimal_places=1,default=100,)
+    max_marks = models.PositiveIntegerField(default=100)
     file_linked = models.FileField(upload_to='class/assignment', null = True, max_length = 1500000,validators=[validate_file_extension])
     classroom = models.ForeignKey(Classroom,on_delete=models.CASCADE,related_name='assignment')
     def __str__(self):
@@ -27,7 +27,7 @@ class Assignment(models.Model):
         
 class AnswerSheet(models.Model):
     file_linked = models.FileField(upload_to="class/answers", null=True, max_length= 1500000,validators=[validate_file_extension])
-    marks_scored = models.DecimalField(max_digits=5,decimal_places=1,default=0)
+    marks_scored = models.PositiveIntegerField(default=0)
     late_submitted = models.BooleanField(default=False)
     checked = models.BooleanField(default=False)
     assignment = models.ForeignKey(Assignment,on_delete=models.CASCADE,related_name='answersheet')
